pyproject.toml
[project]
name = "gemma3n-trial"
version = "0.1.0"
description = "A trial repo for gemma3n"
authors = [
    {name = "Shubham Kumar Patel",email = "shubhkrpatel@gmail.com"}
]
license = {text = "MIT"}
readme = "README.md"
requires-python = ">=3.12,<4.0"
dependencies = [
    "langchain (>=0.3.26,<0.4.0)",
    "langchain-google-genai (>=2.1.8,<3.0.0)",
    "google-genai (>=1.26.0,<2.0.0)",
    "tavily-python (>=0.7.10,<0.8.0)",
    "pydantic (>=2.11.7,<3.0.0)",
    "transformers (>=4.53.2,<5.0.0)",
    "torch (>=2.7.1,<3.0.0)",
    "langgraph (>=0.5.3,<0.6.0)",
    "python-dotenv (>=1.1.1,<2.0.0)",
    "langchain-community (>=0.3.27,<0.4.0)",
    "langchain-tavily (>=0.2.10,<0.3.0)",
    "langchain-groq (>=0.3.6,<0.4.0)",
    "requests (>=2.32.4,<3.0.0)"
]

packages = [
    { include = "gemma3n_trial", from = "src" }
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
